# fix(cluster): optimize the performance for cluster mode