--- a/src/query/service/src/api/rpc/flight_scatter_hash_v2.rs
+++ b/src/query/service/src/api/rpc/flight_scatter_hash_v2.rs
@@ -43,7 +43,10 @@ impl HashFlightScatterV2 {
         func_ctx: FunctionContext,
         scalars: Vec<PhysicalScalar>,
         scatter_size: usize,
-    ) -> Result<Self> {
+    ) -> Result<Box<dyn FlightScatter>> {
+        if scalars.len() == 1 {
+            return OneHashKeyFlightScatter::try_create(func_ctx, &scalars[0], scatter_size);
+        }
         let hash_keys = scalars
             .iter()
             .map(Evaluator::eval_physical_scalar)
@@ -53,12 +56,12 @@ impl HashFlightScatterV2 {
             .map(|k| FunctionFactory::instance().get("sipHash", &[&k.data_type()]))
             .collect::<Result<Vec<_>>>()?;
 
-        Ok(Self {
+        Ok(Box::new(Self {
             func_ctx,
             hash_keys,
             hash_functions,
             scatter_size,
-        })
+        }))
     }
 
     pub fn combine_hash_keys(
@@ -111,12 +114,55 @@ impl HashFlightScatterV2 {
     }
 }
 
+#[derive(Clone)]
+struct OneHashKeyFlightScatter {
+    scatter_size: usize,
+    func_ctx: FunctionContext,
+    indices_scalar: EvalNode<ColumnID>,
+}
+
+impl OneHashKeyFlightScatter {
+    pub fn try_create(
+        func_ctx: FunctionContext,
+        scalar: &PhysicalScalar,
+        scatter_size: usize,
+    ) -> Result<Box<dyn FlightScatter>> {
+        let hash_key = Evaluator::eval_physical_scalar(&scalar)?;
+
+        Ok(Box::new(OneHashKeyFlightScatter {
+            scatter_size,
+            func_ctx,
+            indices_scalar: EvalNode::<ColumnID>::Function {
+                args: vec![
+                    EvalNode::Function {
+                        args: vec![hash_key],
+                        func: FunctionFactory::instance().get("sipHash", &[&scalar.data_type()])?,
+                    },
+                    EvalNode::Constant {
+                        value: DataValue::UInt64(scatter_size as u64),
+                        data_type: DataTypeImpl::UInt64(UInt64Type::new()),
+                    },
+                ],
+                func: FunctionFactory::instance().get("modulo", &[
+                    &DataTypeImpl::UInt64(UInt64Type::new()),
+                    &DataTypeImpl::UInt64(UInt64Type::new()),
+                ])?,
+            },
+        }))
+    }
+}
+
+impl FlightScatter for OneHashKeyFlightScatter {
+    fn execute(&self, data_block: &DataBlock, _num: usize) -> Result<Vec<DataBlock>> {
+        let indices = self.indices_scalar.eval(&self.func_ctx, data_block)?;
+        let col: &PrimitiveColumn<u64> = Series::check_get(indices.vector())?;
+        let indices: Vec<usize> = col.iter().map(|c| *c as usize).collect();
+        DataBlock::scatter_block(data_block, &indices, self.scatter_size)
+    }
+}
+
 impl FlightScatter for HashFlightScatterV2 {
-    fn execute(
-        &self,
-        data_block: &DataBlock,
-        _num: usize,
-    ) -> common_exception::Result<Vec<DataBlock>> {
+    fn execute(&self, data_block: &DataBlock, _num: usize) -> Result<Vec<DataBlock>> {
         let hash_keys = self
             .hash_keys
             .iter()