I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

fix(cluster): optimize the performance for cluster mode(~20%)

before:
```
SELECT COUNT() FROM numbers(1000000000) GROUP BY number % 3;
+-----------+
| COUNT()   |
+-----------+
| 333333333 |
| 333333333 |
| 333333334 |
+-----------+
3 rows in set (6.55 sec)
Read 1000000000 rows, 7.45 GiB in 6.535 sec., 153.02 million rows/sec., 1.14 GiB/sec.
```
after:
```
mysql> SELECT COUNT() FROM numbers(1000000000) GROUP BY number % 3;
+-----------+
| COUNT()   |
+-----------+
| 333333333 |
| 333333334 |
| 333333333 |
+-----------+
3 rows in set (5.08 sec)
Read 1000000000 rows, 7.45 GiB in 5.077 sec., 196.96 million rows/sec., 1.47 GiB/sec.
```

part #7153