@zhang2014 2022-08-29
> Just `~20%` ?

Only on my PC. we need perf test.